% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DBL.R
\name{DBL}
\alias{DBL}
\title{Estimate effects of exposures and identify important ones using debiased lasso.}
\usage{
DBL(Y, x, C, xnew, q = 0.2)
}
\arguments{
\item{Y}{The n by 1 outcome to be analyzed.}

\item{x}{An n by p matrix of exposures.}

\item{C}{An n by q matrix of additional covariates to adjust for.}

\item{xnew}{A design matrix with p exposure values for estimation purpose.}

\item{q}{The false discovery rate threshold.}
}
\value{
A list of values with important main effects and interactions, estimates, 95\% CI, debiased lasso estimates of the coefficients and computation time.
}
\description{
This function estimates the effects of the exposures using debiased lasso 
and identifies the important exposures and their interactions adjusting for confounders while
controlling FDR at a given threshold.
}
