% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knockoffsFull.R
\name{knockoffsFull}
\alias{knockoffsFull}
\title{Identify important exposures and estimate their effects using knockoffs.}
\usage{
knockoffsFull(Y, x, C, xnew, q = 0.2)
}
\arguments{
\item{Y}{The n by 1 outcome to be analyzed.}

\item{x}{An n by p matrix of exposures.}

\item{C}{An n by q matrix of additional covariates to adjust for.}

\item{xnew}{A design matrix with p exposure values for estimation purpose.}

\item{q}{The false discovery rate threshold.}
}
\value{
A list of values with important main effects and interactions, estimates, 95\% CI and computation time.
}
\description{
This function identifies important exposures and their interactions and 
estimates their effects adjusting for confounders while also controlling FDR 
at a given threshold. The method is based on the knockoffs approach.
}
